every rule could only be matched once
Before this commit, if a hyperscan rule matched multiple times for
example 4 times, this single rule would be process 4 times for this
packet, it is difficult for packet processors to distinguish.
So after this commit every rule would be only processed once, and
the from to field would store the max matching range of the pattern.